+++ UlalaAPI/Controllers/ClasseController.cs
@@ -1,8 +1,10 @@
 ﻿using DAL.Entities;
 using DAL.Services;
 using System.Collections.Generic;
+using System;
 using System.Linq;
 using System.Web.Http;
+using UlalaAPI.Helper;
 using UlalaAPI.Mapper;
 using UlalaAPI.Models;
 
@@ -19,12 +21,16 @@ namespace UlalaAPI.Controllers
         /// <param name="Classe">Classe à insérer</param>
         public IHttpActionResult Post(ClasseModel Classe)
         {
-            if (Classe == null || Classe.NomEN == null || Classe.NomFR == null) return BadRequest();
-            else
+            if ((new[] { "Admin" }).Contains(ValidateTokenAndRole.ValidateAndGetRole(Request), StringComparer.OrdinalIgnoreCase))
             {
-                repo.Create(Classe.MapTo<ClasseEntity>());
-                return Ok();
+                if (Classe == null || Classe.NomEN == null || Classe.NomFR == null) return BadRequest();
+                else
+                {
+                    repo.Create(Classe.MapTo<ClasseEntity>());
+                    return Ok();
+                }
             }
+            else return Unauthorized();
         }
         #endregion
 
@@ -35,9 +41,13 @@ namespace UlalaAPI.Controllers
         /// <returns>Liste de toutes les Classes</returns>
         public IHttpActionResult Get()
         {
-            IEnumerable<ClasseModel> Liste = repo.GetAll().Select(Classe => Classe?.MapTo<ClasseModel>());
-            if (Liste.Count() == 0) return NotFound();
-            else return Json(Liste);
+            if ((new[] { "Admin", "User", "Anonyme" }).Contains(ValidateTokenAndRole.ValidateAndGetRole(Request), StringComparer.OrdinalIgnoreCase))
+            {
+                IEnumerable<ClasseModel> Liste = repo.GetAll().Select(Classe => Classe?.MapTo<ClasseModel>());
+                if (Liste.Count() == 0) return NotFound();
+                else return Json(Liste);
+            }
+            else return Unauthorized();
         }
         #endregion
 
@@ -49,9 +59,13 @@ namespace UlalaAPI.Controllers
         /// <returns>Classe avec l'id correspondant</returns>
         public IHttpActionResult Get(int id)
         {
-            ClasseModel Objet = repo.GetOne(id)?.MapTo<ClasseModel>();
-            if (Objet == null) return NotFound();
-            else return Json(Objet);
+            if ((new[] { "Admin", "User", "Anonyme" }).Contains(ValidateTokenAndRole.ValidateAndGetRole(Request), StringComparer.OrdinalIgnoreCase))
+            {
+                ClasseModel Objet = repo.GetOne(id)?.MapTo<ClasseModel>();
+                if (Objet == null) return NotFound();
+                else return Json(Objet);
+            }
+            else return Unauthorized();
         }
         #endregion
 
@@ -62,12 +76,16 @@ namespace UlalaAPI.Controllers
         /// <param name="id">id de la Classe à supprimer</param>
         public IHttpActionResult Delete(int id)
         {
-            if (repo.GetOne(id) == null) return NotFound();
-            else
+            if ((new[] { "Admin" }).Contains(ValidateTokenAndRole.ValidateAndGetRole(Request), StringComparer.OrdinalIgnoreCase))
             {
-                repo.Delete(id);
-                return Ok();
+                if (repo.GetOne(id) == null) return NotFound();
+                else
+                {
+                    repo.Delete(id);
+                    return Ok();
+                }
             }
+            else return Unauthorized();
         }
         #endregion
 
@@ -79,13 +97,17 @@ namespace UlalaAPI.Controllers
         /// <param name="id">Id de la Classe à modifier</param>
         public IHttpActionResult Put(int id, ClasseModel Classe)
         {
-            if (repo.GetOne(id) == null) return NotFound();
-            if (Classe == null || Classe.NomEN == null || Classe.NomFR == null) return BadRequest();
-            else
+            if ((new[] { "Admin" }).Contains(ValidateTokenAndRole.ValidateAndGetRole(Request), StringComparer.OrdinalIgnoreCase))
             {
-                repo.Update(id, Classe.MapTo<ClasseEntity>());
-                return Ok();
+                if (repo.GetOne(id) == null) return NotFound();
+                if (Classe == null || Classe.NomEN == null || Classe.NomFR == null) return BadRequest();
+                else
+                {
+                    repo.Update(id, Classe.MapTo<ClasseEntity>());
+                    return Ok();
+                }
             }
+            else return Unauthorized();
         }
         #endregion
     }
